requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
Django==3.2.6
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
mypy-extensions==0.4.3
numpy==1.21.1
pandas==1.3.1
pathspec==0.9.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.3
six==1.16.0
sqlparse==0.4.1
tomli==1.2.1
